requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
tavily-python>=0.5.0
beautifulsoup4>=4.12.0
requests>=2.32.0
python-dotenv>=1.0.0
lxml>=5.0.0
python-pptx>=1.0.0
pypdf>=4.0.0
